bot/requirements.txt
aiogram==3.13.0
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
asyncpg==0.29.0
redis==5.0.0
sqlalchemy==2.0.0
alembic==1.13.0